+++ common/src/main/java/com/tristankechlo/additionalredstone/init/ModBlocks.java
@@ -16,6 +16,7 @@ public final class ModBlocks {
     public static final RegistryObject<Block> CIRCUIT_MAKER_BLOCK = BLOCKS.register("circuit_maker", CircuitMakerBlock::new);
     public static final RegistryObject<Block> OSCILLATOR_BLOCK = BLOCKS.register("oscillator", OscillatorBlock::new);
     public static final RegistryObject<Block> TIMER_BLOCK = BLOCKS.register("timer", TimerBlock::new);
+    public static final RegistryObject<Block> DIMMABLE_REDSTONE_LAMP = BLOCKS.register("dimmable_redstone_lamp", DimmableRedstoneLampBlock::new);
 
     public static final RegistryObject<Block> AND_GATE_BLOCK = BLOCKS.register("and_gate", () -> new ThreeInputLogicGate(ThreeInputLogic::and));
     public static final RegistryObject<Block> NAND_GATE_BLOCK = BLOCKS.register("nand_gate", () -> new ThreeInputLogicGate(ThreeInputLogic::nand));