+++ forge/src/main/java/com/tristankechlo/additionalredstone/init/ForgeItemGroup.java
@@ -18,23 +18,7 @@ public final class ForgeItemGroup extends CreativeModeTab {
 
     @Override
     public void fillItemList(NonNullList<ItemStack> output) {
-        output.add(ModItems.CIRCUIT_MAKER_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.CIRCUIT_BASE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.NOT_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.AND_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.NAND_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.OR_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.NOR_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.XOR_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.XNOR_GATE_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.T_FLIP_FLOP_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.TOGGLE_LATCH_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.SR_LATCH_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.RS_LATCH_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.SEQUENCER_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.TIMER_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.OSCILLATOR_BLOCK_ITEM.get().getDefaultInstance());
-        output.add(ModItems.SUPERGATE_BLOCK_ITEM.get().getDefaultInstance());
+        AdditionalRedstone.fillItemGroup(output);
         super.fillItemList(output);
     }