+++ Changelog.md
@@ -19,6 +19,9 @@
 - added new block `supergate`
     - special logic gate which can be configured
     - output depends on the configuration for all input combinations
+- added new block `dimmable_redstone_lamp`
+    - emitted light level depends on the strength on the input redstone signal
+    - strongest signal will be used
 
 ### Version 1.18.2 - 1.1.0